Cast tl.program_id to int64 in llama4_rope and qwen2vl_mrope
Both kernels index the flattened batch*seq token dimension with a raw
tl.program_id(0), which Triton materializes as int32. The resulting pointer
arithmetic (base_offset * q_row_stride in llama4_rope, pid * (n_qh * hd) in
qwen2vl_mrope) overflows once the q tensor exceeds 2**31 elements, which is
the same out-of-bounds class that #804 fixed in rope.py and rms_norm.py.

Applies the identical .to(tl.int64) widening. llama4_rope.py was added after
#804 as a fresh kernel body rather than reusing _triton_rope, so it never
inherited the fix, and qwen2vl_mrope.py predates #804 entirely. The Ascend
port of qwen2vl_mrope has the same gap, widened at the pid used for
addressing so loop control stays int32.

Fixes #1335

+++ src/liger_kernel/ops/llama4_rope.py
@@ -39,7 +39,7 @@ def _llama4_rope_kernel(
     Grid: (batch*seq, head)
     """
     # 2D grid
-    pid_bs = tl.program_id(0)  # over batch*seq
+    pid_bs = tl.program_id(0).to(tl.int64)  # over batch*seq
     pid_h = tl.program_id(1)  # over heads
 
     batch_idx = pid_bs // seq_len
